+++ src/Lakona.Game.Server.Hotfix/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Lakona.Game.Server.Hotfix.Tests")]
+[assembly: InternalsVisibleTo("Lakona.Game.Server")]
 [assembly: InternalsVisibleTo("Lakona.Game.Server.Tests")]